backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangorestframework-simplejwt[crypto]==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
django-allauth==0.57.0
dj-rest-auth==5.0.2
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
Pillow==10.0.1
google-generativeai==0.3.0
stripe==7.8.0
requests==2.31.0
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0